Physics now controlled by Chipmunk2D. (#34)
Awaiting nimble to accept our new `chipmunk` package:

https://github.com/nim-lang/packages/pull/2023

The original fork (https://github.com/oprypin/nim-chipmunk) is 5 years old,
and does not work with the latest chipmunk release.

http://chipmunk-physics.net/release/ChipmunkLatest-Docs/

---

- [x] Use a fixed timestep per chipmunk's recommendation
- ~~[ ] Use meters instead of pixels for units (#35)~~
- [x] Fix rotation
- [x] Fix animation with setter procs
- [x] Fix scaling
- [x] Remove `loPhysics` flag

Reviewed-on: https://git.einheit.tech/EinheitTechnologies/shade/pulls/34
Co-committed-by: Avahe Kellenberger <prestige@noreply.%(DOMAIN)s>

+++ tests/shade/circle_test.nim
@@ -2,13 +2,13 @@ import shade
 
 describe "Circle tests":
 
-  let circle = newCircle(vec2(124, 256), 50)
+  let circle = newCircle(dvec2(124, 256), 50)
 
   describe "project":
 
     it "projects the circle correctly onto a horizontal axis":
       let
-        axis = vec2(1, 0)
+        axis = dvec2(1, 0)
         projection = circle.project(VEC2_ZERO, axis)
       # circle.center.x - circle.radius
       assertEquals(projection.x, 74)
@@ -17,7 +17,7 @@ describe "Circle tests":
 
     it "projects the circle correctly onto a vertical axis":
       let
-        axis = vec2(0, 1)
+        axis = dvec2(0, 1)
         projection = circle.project(VEC2_ZERO, axis)
       # circle.center.y - circle.radius
       assertEquals(projection.x, 206)
